requirements.txt
streamlit
pandas
pymupdf
pillow
pytesseract
plotly
reportlab